+++ phpBB2/admin/admin_db_maintenance.php
@@ -442,8 +442,9 @@ switch($mode_id)
 				else // anonymous user does not exist
 				{
 					// Recreate entry
-					$sql = "INSERT INTO " . USERS_TABLE . " (user_id, username, user_level, user_regdate, user_password, user_email, user_icq, user_website, user_occ, user_from, user_interests, user_sig, user_viewemail, user_style, user_aim, user_yim, user_msnm, user_posts, user_attachsig, user_allowsmile, user_allowhtml, user_allowbbcode, user_allow_pm, user_notify_pm, user_allow_viewonline, user_rank, user_avatar, user_lang, user_timezone, user_dateformat, user_actkey, user_newpasswd, user_notify, user_active)
-						SELECT " . ANONYMOUS . ", 'Anonymous', 0, 0, '', '', '', '', '', '', '', '', 0, NULL, '', '', '', 0, 0, 1, 1, 1, 0, 1, 1, 0, '', '', 0, '', '', '', 0, 0 WHERE NOT EXISTS (SELECT 1 FROM (SELECT DISTINCT user_id FROM " . USERS_TABLE . ") guest_current WHERE user_id = " . ANONYMOUS . ")";
+					$guest_profile_parts = phpbb_profile_guest_insert_parts($db);
+					$sql = "INSERT INTO " . USERS_TABLE . " (user_id, username, user_level, user_regdate, user_password, user_email, user_icq, user_website, user_occ, user_from, user_interests, user_sig, user_viewemail, user_style, user_aim, user_yim, user_msnm, user_posts, user_attachsig, user_allowsmile, user_allowhtml, user_allowbbcode, user_allow_pm, user_notify_pm, user_allow_viewonline, user_rank, user_avatar, user_lang, user_timezone, user_dateformat, user_actkey, user_newpasswd, user_notify, user_active" . $guest_profile_parts[0] . ")
+						SELECT " . ANONYMOUS . ", 'Anonymous', 0, 0, '', '', '', '', '', '', '', '', 0, NULL, '', '', '', 0, 0, 1, 1, 1, 0, 1, 1, 0, '', '', 0, '', '', '', 0, 0" . $guest_profile_parts[1] . " WHERE NOT EXISTS (SELECT 1 FROM (SELECT DISTINCT user_id FROM " . USERS_TABLE . ") guest_current WHERE user_id = " . ANONYMOUS . ")";
 					$result = $db->sql_write($sql);
 					if ( !$result )
 					{